+++ docs/CHANGELOG.rst
@@ -15,6 +15,8 @@ Added
 -----
 - Allow creating multiple ``Variant`` objects with one API request
 - Allow creating, updating and deleting ``VariantAnnotation`` objects via API
+- Add ``VariantAnnotationTranscript`` endpoint
+
 Changed
 -------